+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Extracted shared Node.js version types into `nodejs-data` crate. ([#1328](https://github.com/heroku/buildpacks-nodejs/pull/1328))
+- Node.js inventory now includes release schedule metadata. ([#1330](https://github.com/heroku/buildpacks-nodejs/pull/1330))
 
 ## [5.5.4] - 2026-03-12